* move DebugManager to another package
